+++ eras/alonzo/impl/test/Test/Cardano/Ledger/Alonzo/BinarySpec.hs
@@ -4,7 +4,6 @@ module Test.Cardano.Ledger.Alonzo.BinarySpec (spec) where
 
 import Cardano.Ledger.Alonzo
 import Cardano.Ledger.Alonzo.Genesis
-import Cardano.Ledger.Alonzo.Scripts
 import Test.Cardano.Ledger.Alonzo.Arbitrary ()
 import Test.Cardano.Ledger.Alonzo.Binary.RoundTrip (roundTripAlonzoCommonSpec)
 import Test.Cardano.Ledger.Alonzo.TreeDiff ()
@@ -23,6 +22,3 @@ spec = do
     roundTripAlonzoCommonSpec @Alonzo
     -- AlonzoGenesis only makes sense in Alonzo era
     roundTripEraSpec @Alonzo @AlonzoGenesis
-    -- CostModel serialization changes drastically for Conway, which requires a different
-    -- QuickCheck generator, hence Arbitrary can't be reused
-    roundTripEraSpec @Alonzo @CostModels