+++ eras/conway/impl/src/Cardano/Ledger/Conway/PParams.hs
@@ -112,8 +112,7 @@ import Cardano.Ledger.HKD (
  )
 import Cardano.Ledger.Plutus.CostModels (
   CostModel,
-  decodeCostModelFailHard,
-  decodeValidAndUnknownCostModels,
+  decodeCostModel,
   encodeCostModel,
   mkCostModel,
   mkCostModels,
@@ -674,7 +673,7 @@ instance DecCBOR (UpgradeConwayPParams Identity) where
         <! From
         <! From
         <! From
-        <! D (decodeCostModelFailHard PlutusV3)
+        <! D (decodeCostModel PlutusV3)
 
 instance Crypto c => EraPParams (ConwayEra c) where
   type PParamsHKD f (ConwayEra c) = ConwayPParams f (ConwayEra c)
@@ -1055,7 +1054,7 @@ updateField = \case
   11 -> field (\x up -> up {cppTau = THKD (SJust x)}) From
   16 -> field (\x up -> up {cppMinPoolCost = THKD (SJust x)}) From
   17 -> field (\x up -> up {cppCoinsPerUTxOByte = THKD (SJust x)}) From
-  18 -> field (\x up -> up {cppCostModels = THKD (SJust x)}) $ D decodeValidAndUnknownCostModels
+  18 -> field (\x up -> up {cppCostModels = THKD (SJust x)}) From
   19 -> field (\x up -> up {cppPrices = THKD (SJust x)}) From
   20 -> field (\x up -> up {cppMaxTxExUnits = THKD (SJust x)}) From
   21 -> field (\x up -> up {cppMaxBlockExUnits = THKD (SJust x)}) From